requirements.txt
Django>=4.2.0,<5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter
gunicorn>=21.2.0  # Production server
whitenoise>=6.5.0  # Static file serving
django-environ>=0.11.2
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
python-decouple>=3.8
django-widget-tweaks>=1.5.0
